Changed dblayer.version:
Deleted the field cdmelements from the volumes table.
It was a serialization of an array, thus causing problems when
changing the verison of java.
It is now built at server statup when adding a volume in cache.

Fixed a problem with the AdvancedSearch.java when selecting dictionaries instead of volumes. All the volumes of all the dictionaries where selected.



git-svn-id: https://svn.ligforge.imag.fr/svn/jibiki/trunk@1015 73f06212-78d2-4448-adb7-01a149c5d412

Former-commit-id: 77468d491072c5ae864de683d1aa1f5cc825f1d8